decodebin opens what the SDK writes, and the Windows path compiles
An integration test behind the gst feature writes one second of colour
bars with the pure Rust Matroska writer and reads it back through
decodebin: I420 at canvas caps, every frame accounted for, and the
audio track at 48 kHz stereo. The Unix only imports in the GStreamer
module are cfg gated, checked by building the crate with unix forced
off, which is the cfg set a windows-msvc build sees.

+++ crates/godwinmix-sdk/src/media/matroska.rs
@@ -526,7 +526,7 @@ mod tests {
             while *at < bytes.len() {
                 let id = read_id(bytes, at);
                 let size = read_size_field(bytes, at);
-                let known = size.unwrap_or(u64::MAX as u64);
+                let known = size.unwrap_or(u64::MAX);
                 match id.as_slice() {
                     ID_SEGMENT => continue, // unknown size: its children follow
                     ID_CLUSTER => {
